+++ CSharp/Concentus/Celt/Structs/CELTDecoder.cs
@@ -558,7 +558,7 @@ namespace Concentus.Celt.Structs
                 // If no entropy decoder was passed into this function, we need to create
                 // a new one here for local use only. It only exists in this function scope.
                 dec = new EntropyCoder();
-                dec.dec_init(data, data_ptr, (uint)len);
+                dec.dec_init(data, data_ptr, len);
             }
 
             if (C == 1)
